backend/requirements.txt
# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Database
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1

# Data Validation
pydantic==2.6.1
pydantic[email]==2.6.1
pydantic-settings==2.1.0

# Authentication & Security
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9
cryptography==42.0.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# Cache & Storage
redis==5.0.1
aiofiles==23.2.1

# Cloud Storage (Alibaba Cloud OSS)
oss2==2.18.5

# Environment
python-dotenv==1.0.0

# System Monitoring
psutil==5.9.8
docker==7.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
